[IMP] pms_fastapi: expose reservation summary in folio sale lines
